Take labels from `flow.run_config` if present
If a flow has a `run_config`, the labels on the `run_config` are used,
and `flow.environment` is ignored. This updates the label logic in the
UI to take labels from the `run_config` if present.

Also deletes some dead code left in the `Details-Tile.vue` leftover from
moving label editing to the `LabelEdit` component.