Make request credentials forwarding customizable
When running Prefect on premise, it's sometimes necessary for the UI to
forward authentication cookies (e.g. when the authentication layer is provided
by a reverse proxy and not the Prefect server). This change allows for the UI
to be customized by providing an extra env variable named
`VUE_APP_REQUEST_CREDENTIALS_MODE`, which will cause the UI to forward cookies
to the Apollo server if the variable is set with `include` as its value. If the
variable is left unset, the behaviour is the same as before.